+++ packages/web/hooks/window/use-window-scroll.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect } from "react";
+import { useEffect, useState } from "react";
 
 /**
  * Hook into the window scroll position.
